added payment

+++ backend/src/main/java/com/MHunter/mhunter/model/Invoice.java
@@ -23,7 +23,7 @@ public class Invoice {
     private Double transportFee;
     private Double others;
     private String paymentType;
-    private Double paidAmount;
+    private Long paidAmount;
     private Double artistFee;
     private Double discount;
